pubsub-negative-analyzer/requirements.txt
google-cloud-language==2.5.0
google-cloud-secret-manager==2.10.0
requests==2.28.1